Add XMLHttpRequest interception
This commit adds XHR interception both the IWindow for xhr constructor
and also explicitly to capture some of the methods and attributes of the
XHR class.
Added some initial tests, but the xhr provided with jsdom is not fully
working. Need to mainly test with the browser. But at least it is clear
that the interception is working as expected.