Support intercepting event listeners
There are two ways to add event handlers, one useing
`addEventListener('evant', ...)` and the other one is directly assigning
to the `onevent = ...`.
While the latter is only a callback function, the former accepts both a
function and a special form an object (with `handleEvent` method).

This code adds support for all of these caes in the browser. This will
allow us to intercept the execution of the event listeners themselves.
We will use this later to support flowlets.