Improve bundle size
When an exported value from a module is unused, Rollup will remove the
assignment but leave the initialization in the code for another tool
(e.g. Google Closure) to optimize.
For intercepted properties of the DOM, we endup with a bunch of unused
intercept* methods in the code.
This commit extends the generateBundle callback to comments these unused
calls.
That will make the bundle smaller and more performant.