Prevent duplicate interception
When intercepting a class hierarchy, it may not be clear which class in
the chain has implemented a method/attribute/constructor.
If someone add interceptors for both parent and child class,
we should still expect the intercepion work properly
and we have only a single interceptor per method.

In this commit, the intercept* functions will first check if a
corresponding interceptor was already installed and return that,
otherwise it would instanciate a new one.