Fixed logic for removeEventListener in flowlet
The flowlet logic for handleing event listeners was converting a callback to its intercepted version when `addEventListener` was called before passing it to the DOM api. The logic for `removeEventListener` was supposed to do the same to ensure DOM api always recieves the same listner. However, it was passing the original function.

Added jest test for capturing this as well.