Added a utility to automatically extend react component props.
In some cases, we would want to add certain info to all react component
(classical or functional). This utilitiy uses the interception layers to
achieve that.

Test is done in the test react app in a simple way, need to add better
jest tests. Will do once Surface and flowlet is added.