Enable passing and updateing the Flowlets in React methods.
This logic uses the props extension logic to allow each props of the
react components to carry a flowlet value and then pop that back on
stack so it can be further passed around correctly.
(Note, this commit only handles render related methods, there will be
another on handling other react hooks as well.