Enable a surface renderer fall back.
Now that initialization of surface and use of it (via
`getSurfaceRenderer` ) is separated from each other, we can simplify the
usage by allowing the function to accept a fallback instead of throwing
exception.

This way, applications can easily init the whole of AutoLogging
conditionally in one place, and then use the surface regardless of
initilozation (i.e. use the new surface if initialized, else use a
simple default).