Replace flowlet marker with a more generic approch
Instead of only adding the flowlet to the request, introduced a new init
option where app can send a function to be used for marking every
request that goes out.

With this approach, and app can optionally use a `Hook` and further
expand to possibility to individual parts of the its code to inspect the
requests and append any parameter to them.

tested with fetch and xhr requests in the react-testapp