Extend Channel to be pipeable
Sometimes we want to chain (pipe) channels together so that first all
the event listeners of first channel recieve the events, and then next
channel recieve it.

This is usually useful for either scheduling the second channel for
later, or for transforming the events between channels.

This commit adds such capability to channels.

Also, separted Hook and Channel into their own bundle so that outside
applications can use them independently.