A new algorithm to track flowlets in react
Two sets of changes are included in this commit:
1- a set of common hooks that have a function callback are intercepted
in react. This set is then used to ensure those callbacks carry the
flowlet of the render function forward.
2- In each render function, instead of relying on the prop extensions to
detect the current flowlet, we use the the SurfaceContext to use the
flowlet of the current surface. This flowlet is then picked up by other
lifecycle methods or hooks and propagated.

I also updated the test app to ensure the full flowlet initialization is
applied for all async browser api.

Together, this new algorithm give much bigger coverage for flowlets.