Avoid double interception of global primordials (Promise, ...) & fix bundle
While in a normal application we have guarantee that a module loads
once, in test environment the same module may be loaded multiple times.
Since we intercept and change the primordial objects such as Promise, we
might see the effect of interception across tests.

The code fix here ensure that we don't create extra wrappers for such
objects. The asserts in the code did catch such issue.

Also changed the bundle config so that utils classes remain in
hyperionUtils bundle.