Added cleanup task for FlowletManager
When FlowletManager.pop is called, if the passed flowlet is not the same
as the .top() we don't really need to immediately try to remove it from
the stack. If the stack get very large, searching for such flowlets in
the stack can take significant time.

Instead in the new algorithm, these cases are added to a separate set to
be cleaned up later.

Also, in case something is left in the stack for a long time, when
browser is idle, we expect stack to be empty for then, and if not, we
just cleanup to release all those objects.

This situation can happen if for whatever reason (e.g. a code exception,
react suspense, ...) application does not run the corresponding .pop
method.

Also updated the .onPush and .onPop to pass more information (e.g.
replacementFlowlet, popSucceeded).