Only include pathname for network trigger flowlets.
To ensure the triggerName for fetch/xhr is short and also does not
include any sensitive information, changed the code to only keep the
pathname in the flowlet name.

Now that we need to part the URL, added the result in the RequestInfo so
we can reuse it in other places. For now, won't expose it externally,
but can do it in a separate commit.