Fix proxy container usage for Surface during strict mode
React in strict mode (dev time), renders components twice. That means
portal containers my be empty the first round, and have some content the
second round. We had a conditional usage of such container for Surface
wrappers that would fail the second time.

This change allows us to check if the container is empty or has the same
attribute as before, and hence avoid adding the Surface wrapper.