Added a simple 'mark' api to FlowletManager
Sometimes, we want to make sure when a function is called, it 'forks'
the top flow and marks its execution for the rest of the stack.
In these case, we should not use the FlowletManager.wrap since it
"preserves" the existing flowlet for a callback.
The `FlowletManager.mark` function will return a new function that
automatically adds the markings.