Add pageload trigger
When page is loading we don't have any trigger yet applied. Now that we have a root for all flowlets, we can initially set a pageload trigger for all things that happen at the begining.

Also, we ran into a new root flowlet that should be added to the active roots, we an just add the last-seen trigger for that flowlet.