Reduce the volume of element value change events.
React keeps setting the value of the input elements behind the scene
(even on mouse move!) and
since we have an interception on the setter, we keep emitting the
corresponding event.

This commit checks the new value of element.checked and emit event only
if it has chaned.

Also, as an optimization, if the Surface does not have tracking
capability, then we know that all inputs under that surface will be
ignored. so, added an explicit code to filter such cases out.