Implement a component_visibility event
This commits adds support for tracking the actual visibility of the
Surfaces. This uses `intersectionObserver` and can be configured to
track any desired threshold of the Surface visibility compared to the
root component.
The event also carries the visibility area and parent view port size for
when we want to know where the surface is rendered on the screen or how
much area it is covering.

See the following for more info:
https://developer.mozilla.org/en-US/docs/Web/API/IntersectionObserverEntry

Note: still need to decide how to handle 1) SurfaceProxy and 2) Surface
with more than one child element.