Fixed visiblity tracking with SurfaceProxy
I noticed we don't pass the `capability` correctly for <SurfaceProxy>
components.
After fixing that, I realized it is possible that the mount event for
surface proxy fires before the mount event for the surface itself. This
does not cause issue for mount event itself, but breaks the visibility
algorithm.
So, I changed the algorithm to keep two sets of info:
* map surface => SurfaceEvent so we can later link the events
* map element => surface

This fixes issues with visibility of surface proxies, also removes the
need for looking up the `element.parentElement`