Making ALGraph more reusable
In order to make it possible for other applications to also instanciate
their own version of the ALGraph, we need to move most of ths
functionality into the ALGraph itself.

This first diff moves the basic clear/fit/snapshot into the ALGraph
class to that they can be directly an easily called from other react
components.