Improve scratch data of nodes and click handling on nodes.
We were not using the scratch data of nodes currently and could
potentially overwrite other data.
In this commit, added a namespace for the scratch data. Also, enable a
generic onClick mechanism that can be based on the field of the scratch
data.
This way, child classes can extend the graph with new data and enjoy the
same click handling.