Enable caching of interactivity detection
In the previous algorith, everytime we get a node, we would walk up the
ancestor chain and perform several comparisions to check if that node is
a good candidate for the sub-root of the interaction tree.
In some cases (e.g. mouseover) this could be very repetitive and
expensive.
Specially cases that involve reading client height and width of elements
and window could be expensive.

The new logic caches the result on the node itself and tries to avoid
recalculation every time. This optimization for now is behind a flag
(`optimizeInteractibiltyCheck`) until we make sure this is really
improving the performance and not causing any other functional issues.